Move std_pool module to pool::standard
