+++ vulkano/src/descriptor_set/persistent.rs
@@ -15,13 +15,13 @@ use crate::descriptor_set::descriptor::DescriptorImageDesc;
 use crate::descriptor_set::descriptor::DescriptorImageDescArray;
 use crate::descriptor_set::descriptor::DescriptorImageDescDimensions;
 use crate::descriptor_set::descriptor::DescriptorType;
+use crate::descriptor_set::pool::standard::StdDescriptorPoolAlloc;
 use crate::descriptor_set::pool::DescriptorPool;
 use crate::descriptor_set::pool::DescriptorPoolAlloc;
 use crate::descriptor_set::DescriptorSet;
 use crate::descriptor_set::DescriptorSetDesc;
 use crate::descriptor_set::DescriptorSetLayout;
 use crate::descriptor_set::DescriptorWrite;
-use crate::descriptor_set::StdDescriptorPoolAlloc;
 use crate::descriptor_set::UnsafeDescriptorSet;
 use crate::device::Device;
 use crate::device::DeviceOwned;
